More Separation into partial definintions for vector
