+++ DataScience/Core/Vector/Vector.cs
@@ -755,222 +755,6 @@ namespace DataScience
         }
 
 
-
-
-
-        public static Vector Diff(Vector vector)
-        {
-            if (vector.Columns > 1)
-            {
-                throw new Exception("Diff is for use with 1D Vectors ONLY");
-            }
-
-            // Ensure there is enough space for all the data
-            long size = vector.MemorySize() * 2;
-
-            uint flag = vector.Id;
-            vector.gpu.AddFlags(flag);
-            vector.gpu.DeCacheLRU(size, true);
-
-            // Make the Output Vector
-            Vector Output = new Vector(vector.gpu, new float[vector.Value.Length - 1], vector.Columns);
-
-            MemoryBuffer<float> buffer = Output.GetBuffer(); // Output
-            MemoryBuffer<float> buffer2 = vector.GetBuffer(); // Input
-
-            vector.gpu.diffKernel(vector.gpu.accelerator.DefaultStream, buffer.Length, buffer.View, buffer2.View);
-
-            vector.gpu.accelerator.Synchronize();
-
-            buffer.CopyTo(Output.Value, 0, 0, Output.Value.Length);
-
-            vector.gpu.RemoveFlags(flag);
-
-            return Output;
-        }
-
-        public void Diff_IP()
-        {
-            Vector Output = Vector.Diff(this);
-
-            this.Dispose();
-            this.Value = Output.Value[..];
-            this.Id = Output.Id;
-
-            return;
-        }
-
-
-        public static float DotProduct(Vector vectorA, Vector vectorB)
-        {
-            return ConsecutiveOP(vectorA, vectorB, Operations.multiplication).Sum();
-        }
-        public static float DotProduct(Vector vectorA, float scalar)
-        {
-            return ConsecutiveOP(vectorA, scalar, Operations.multiplication).Sum();
-        }
-        public float DotProduct(Vector vectorB)
-        {
-            return ConsecutiveOP(this, vectorB, Operations.multiplication).Sum();
-        }
-        public float DotProduct(float scalar)
-        {
-            return ConsecutiveOP(this, scalar, Operations.multiplication).Sum();
-        }
-
-
-        public static Vector Normalise(Vector vectorA)
-        {
-            return ConsecutiveOP(vectorA, 1f / vectorA.Sum(), Operations.multiplication);
-        }
-        public void Normalise_IP()
-        {
-            ConsecutiveOP_IP(1f / this.Sum(), Operations.multiplication);
-        }
-
-
-        /// <summary>
-        /// Takes the absolute value of all values in the Vector.
-        /// IMPORTANT : Use this method for Vectors of Length less than 100,000
-        /// </summary>
-        /// <param name="vector"></param>
-        /// <returns></returns>
-        public static Vector Abs(Vector vector)
-        {
-            Vector vec = vector.Copy();
-            vec.Abs_IP();
-            return vec;
-        }
-        /// <summary>
-        /// Takes the absolute value of all values in this Vector.
-        /// IMPORTANT : Use this method for Vectors of Length less than 100,000
-        /// </summary>
-        public void Abs_IP()
-        {
-            if (this.Value.Min() > 0f)
-            {
-                return;
-            }
-
-            for (int i = 0; i < this.Value.Length; i++)
-            {
-                this.Value[i] = MathF.Abs(this.Value[i]);
-            }
-
-            if (this.Id != 0)
-            {
-                UpdateCache();
-            }
-
-            return;
-        }
-        /// <summary>
-        /// Runs on Accelerator. (GPU : Default)
-        /// Takes the absolute value of all the values in the Vector.
-        /// IMPORTANT : Use this method for Vectors of Length more than 100,000
-        /// </summary>
-        /// <param name="vector"></param>
-        /// <returns></returns>
-        public static Vector AbsX(Vector vector)
-        {
-            Vector vec = vector.Copy();
-            vec.AbsX_IP();
-            return vec;
-        }
-        /// <summary>
-        /// Runs on Accelerator. (GPU : Default)
-        /// Takes the absolute value of all the values in this Vector.
-        /// IMPORTANT : Use this method for Vectors of Length more than 100,000
-        /// </summary>
-        public void AbsX_IP()
-        {
-            // Check if the input & output are in Cache
-            MemoryBuffer<float> buffer = this.GetBuffer(); // IO
-
-            gpu.absKernel(gpu.accelerator.DefaultStream, buffer.Length, buffer.View);
-
-            gpu.accelerator.Synchronize();
-
-            buffer.CopyTo(this.Value, 0, 0, this.Value.Length);
-
-            return;
-        }
-
-
-
-
-
-        /// <summary>
-        /// Determines if All the values in the Vector are Non-Zero
-        /// </summary>
-        /// <returns></returns>
-        public static bool All(Vector vector)
-        {
-            return !vector.Value.Contains(0f);
-        }
-        /// <summary>
-        /// Determines if All the values in this Vector are Non-Zero
-        /// </summary>
-        /// <returns></returns>
-        public bool All()
-        {
-            return !this.Value.Contains(0f);
-        }
-
-
-        public static Vector Reciprocal(Vector vector)
-        {
-            Vector vec = vector.Copy();
-            vec.Reciprocal_IP();
-            return vec;
-        }
-        public void Reciprocal_IP()
-        {
-            // Check if the input & output are in Cache
-            MemoryBuffer<float> buffer = this.GetBuffer(); // IO
-
-            gpu.reciprocalKernel(gpu.accelerator.DefaultStream, buffer.Length, buffer.View);
-
-            gpu.accelerator.Synchronize();
-
-            buffer.CopyTo(this.Value, 0, 0, this.Value.Length);
-
-            return;
-        }
-
-
-        public static Vector Reverse(Vector vector)
-        {
-            return new Vector(vector.gpu, vector.Value.Reverse().ToArray(), vector.Columns);
-        }
-        public void Reverse_IP()
-        {
-            this.Value = this.Value.Reverse().ToArray();
-            if (this.Id != 0) { UpdateCache(); }
-            return;
-        }
-        public static Vector ReverseX(Vector vector)
-        {
-            Vector vec = vector.Copy();
-            vec.ReverseX_IP();
-            return vec;
-        }
-        public void ReverseX_IP()
-        {
-            // Check if the input & output are in Cache
-            MemoryBuffer<float> buffer = this.GetBuffer(); // IO
-
-            gpu.reverseKernel(gpu.accelerator.DefaultStream, buffer.Length >> 1, buffer.View);
-
-            gpu.accelerator.Synchronize();
-
-            buffer.CopyTo(this.Value, 0, 0, this.Value.Length);
-
-            return;
-        }
-
-
-        
        
 
     }
